litecart/requirements.txt
pytest==7.1.2
selenium==4.4.0
webdriver-manager==3.8.3
faker==14.1.0
allure-pytest==2.9.45
mysql == 0.0.3
mysql-connector-python==8.0.30